refactor: norminette compliance pass on lexer/parser
Split oversized functions and files (build_args_array, parser.c),
relocate scan_word/skip_word by responsibility, fix EOL_OPERATOR
and alignment errors. No behavior change.
