--- minishell_debug.dSYM/Contents/Resources/Relocations/x86_64/minishell_debug.yml
@@ -1,5 +1,0 @@
----
-triple:          'x86_64-apple-darwin'
-binary-path:     minishell_debug
-relocations:     []
-...
