+++ src/parser/parser_utils.c
@@ -6,7 +6,7 @@
 /*   By: kkaratsi <kkaratsi@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2026/06/18 14:39:57 by kkaratsi          #+#    #+#             */
-/*   Updated: 2026/06/18 14:39:59 by kkaratsi         ###   ########.fr       */
+/*   Updated: 2026/06/19 13:41:58 by kkaratsi         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
